+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ClientRenderer: Use `<pre>` instead of `<div>` for diagram containers.
 
 ### Added
-- ClientRenderer: Support changing the container tag
-  with the `ContainerTag` option.
+- Support changing the container tag with the `ContainerTag` option.
+  This option is available on ClientRenderer, ServerRenderer, and Extender.
 
 ## [0.3.0] - 2022-12-19
 ### Changed